+++ platform/android/lynx_android/src/main/java/com/lynx/tasm/behavior/ui/list/container/UIListContainer.java
@@ -32,12 +32,10 @@ import com.lynx.tasm.LynxError;
 import com.lynx.tasm.LynxSubErrorCode;
 import com.lynx.tasm.ThreadStrategyForRendering;
 import com.lynx.tasm.base.LLog;
-import com.lynx.tasm.base.TraceEvent;
 import com.lynx.tasm.behavior.LynxContext;
 import com.lynx.tasm.behavior.LynxProp;
 import com.lynx.tasm.behavior.LynxUIMethod;
 import com.lynx.tasm.behavior.LynxUIMethodConstants;
-import com.lynx.tasm.behavior.event.EventTarget;
 import com.lynx.tasm.behavior.ui.IDrawChildHook;
 import com.lynx.tasm.behavior.ui.LynxBaseUI;
 import com.lynx.tasm.behavior.ui.list.ListEventManager;
@@ -50,7 +48,6 @@ import com.lynx.tasm.event.EventsListener;
 import com.lynx.tasm.event.LynxDetailEvent;
 import com.lynx.tasm.event.LynxScrollEvent;
 import com.lynx.tasm.gesture.GestureArenaMember;
-import com.lynx.tasm.gesture.arena.GestureArenaManager;
 import com.lynx.tasm.gesture.detector.GestureDetector;
 import com.lynx.tasm.gesture.handler.BaseGestureHandler;
 import com.lynx.tasm.utils.PixelUtils;
@@ -88,6 +85,7 @@ public class UIListContainer extends UISimpleView<ListContainerView>
   private UIComponent mPrevStickyBottomItem = null;
   private boolean mEnableFadeInAnimation = false;
   private int mUpdateAnimationFadeInDuration = DEFAULT_FADE_IN_ANIMATION_DURATION;
+  private boolean mEnableRecycleStickyItem = true;
   private boolean mEnableScrollEndEvent = false;
   private boolean mEnableScrollStateChangeEvent = false;
   private boolean mEnableBatchRender = false;
@@ -580,6 +578,11 @@ public class UIListContainer extends UISimpleView<ListContainerView>
     mUpdateAnimationFadeInDuration = value;
   }
 
+  @LynxProp(name = "experimental-recycle-sticky-item", defaultBoolean = true)
+  public void setEnableRecycleStickyItem(boolean value) {
+    mEnableRecycleStickyItem = value;
+  }
+
   @LynxProp(name = "need-visible-item-info", defaultBoolean = false)
   public void setNeedVisibleItemInfo(boolean value) {
     mEnableNeedVisibleItemInfo = value;
@@ -965,7 +968,10 @@ public class UIListContainer extends UISimpleView<ListContainerView>
       while (iterator.hasNext()) {
         Map.Entry<Integer, UIComponent> entry = iterator.next();
         if (entry.getValue() == child) {
-          stickyItems.remove(entry.getKey());
+          if (mEnableRecycleStickyItem) {
+            resetStickyItem((UIComponent) child);
+          }
+          iterator.remove();
           break;
         }
       }
@@ -993,6 +999,13 @@ public class UIListContainer extends UISimpleView<ListContainerView>
     }
   }
 
+  private void resetStickyItem(UIComponent component) {
+    if (component.getView() != null) {
+      component.getView().setTranslationY(0);
+      setChildTranslationZ(component);
+    }
+  }
+
   public void updateStickyTops(int offsetTop) {
     if (!mEnableListSticky) {
       return;
@@ -1012,12 +1025,10 @@ public class UIListContainer extends UISimpleView<ListContainerView>
         // cache potential next sticky item
         nextStickyTopItem = top;
         // hold its position
-        top.getView().setTranslationY(0);
-        setChildTranslationZ(top);
+        resetStickyItem(top);
       } else if (stickyTopItem != null) {
         // sticky top item found, hold upper sticky top's position
-        top.getView().setTranslationY(0);
-        setChildTranslationZ(top);
+        resetStickyItem(top);
       } else {
         stickyTopItem = top;
       }
@@ -1064,12 +1075,10 @@ public class UIListContainer extends UISimpleView<ListContainerView>
         // cache potential next sticky item
         nextStickyBottomItem = bottom;
         // hold its position
-        bottom.getView().setTranslationY(0);
-        setChildTranslationZ(bottom);
+        resetStickyItem(bottom);
       } else if (stickyBottomItem != null) {
         // sticky bottom item found, hold upper sticky top's position
-        bottom.getView().setTranslationY(0);
-        setChildTranslationZ(bottom);
+        resetStickyItem(bottom);
       } else {
         stickyBottomItem = bottom;
       }